Moves main files to new package.
